submissions/PrepWise/backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
# agno
openai
firecrawl-py
agno --pre
exa-py